+++ index.html
@@ -229,7 +229,7 @@
   <template id="messages">
     <div class="img-upload__message  img-upload__message--loading">Загружаем...</div>
   </template>
-
-  <script src="main.js"></script>
+  <script src="https://cdn.jsdelivr.net/npm/lodash@4.17.20/lodash.min.js"></script>
+  <script type='module' src="js/main.js"></script>
 </body>
 </html>
